Add page change detection to scraper using hash-based caching
